Build Fix 2

+++ src/components/protocol_handler/include/protocol_handler/protocol_handler_impl.h
@@ -431,10 +431,10 @@ class ProtocolHandlerImpl
                                     const RawMessagePtr &packet);
     RESULT_CODE EncryptData(const int32_t connection_key, const uint8_t service_type,
                             const uint8_t* const data_in, const size_t data_in_size,
-                            const uint8_t **data_out, size_t *data_out_size);
+                            const uint8_t **data_out, uint32_t *data_out_size);
     RESULT_CODE DecryptData(const int32_t connection_key, const uint8_t service_type,
                             const uint8_t* const data_in, const size_t data_in_size,
-                            uint8_t**  data_out, size_t *data_out_size);
+                            uint8_t**  data_out, uint32_t *data_out_size);
 
   private:
     /**
